Windows build chain installs most prebuilts
The dependency build chain builds and installs most
of the libraries for msvcr71 (Python 2.5).

TODO:
Still need to check SDL_ttf and MikMod, as well as
the ffmpeg libraries (no rush on those).

The Pygame unit tests do not complete for Python 2.5.

Get the following periodic Windows console error message:

  This application has requested the Runtime to terminate it in an unusual way.
  Please contact the application's support team for more information.

Build the dependencies for msvcr90.dll. Attempt to get msys_build_deps.py to
build the DLLs directly, either by running the configure scripts with msvcr71
and switching to msvcr90 for Make, or adding manifest files for the configure
test programs.