Move new buffer support determination to setup.py
Allow the ENABLE_NEWBUF C macro to be set in the compiler shell command. This
allows the decision to enable new buffer protocol support to be made in the
build chain, particularly setup.py. Now new buffer protocol support and the
deletion of src/newbuffer.c are controlled by the same piece of code.